requirements.txt
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
billiard==4.2.0
black==24.3.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
config==0.5.1
environ==1.0
exceptiongroup==1.2.0
fastapi==0.110.0
fastapi-cache2==0.2.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
kombu==5.3.5
Mako==1.3.2
MarkupSafe==2.1.5
multipart==0.2.4
mypy-extensions==1.0.0
packaging==24.0
pathlib==1.0.1
pathspec==0.12.1
pendulum==3.0.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pydentic==0.0.1.dev3
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-stdnum==1.20
redis==4.6.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
tesseract==0.1.3
time-machine==2.14.0
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.28.0
vine==5.1.0
wcwidth==0.2.13
gunicorn
flower
#python==3.10.12


#pip freeze > requirements.txt
aioredis==2.0.1


